requirements-f.txt
packaging>=20.9
colorlog>=6.6.0
pyahocorasick>=1.4.2
jsoneditor>=1.4.3
hjson>=3.0.2
numpy>=1.21.4
pandas>=1.2.4
pandarallel>=1.5.2
gensim>=4.1.2
tensorboard>=2.7.0
google-auth>=1.8.0
markdown>=3.3.4
requests-oauthlib>=1.3.0
-f https://download.pytorch.org/whl/torch_stable.html torch>=1.8.2
-f https://download.pytorch.org/whl/torch_stable.html torchvision==0.11.1
tokenizers>=0.12.0
transformers==4.17.0
pytorch-lightning==1.6.0
